Refactor the tests for the distribution version
This includes:
* moving the tests from test_init.py into test_assorted.py
* rely on PYTHON_VERSION variable rather than the nested try/except
* monkeypatch inside the test function and get rid of cerberus fixture
