+++ cerberus/tests/test_assorted.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from decimal import Decimal
+from pkg_resources import Distribution, DistributionNotFound
 
 from pytest import mark
 
@@ -8,6 +9,37 @@ from cerberus import TypeDefinition, Validator
 from cerberus.tests import assert_fail, assert_success
 from cerberus.utils import validator_factory
 from cerberus.validator import BareValidator
+from cerberus.platform import PYTHON_VERSION
+
+
+if PYTHON_VERSION > 3 and PYTHON_VERSION < 3.4:
+    from imp import reload
+elif PYTHON_VERSION >= 3.4:
+    from importlib import reload
+else:
+    pass  # Python 2.x
+
+
+def test_pkgresources_version(monkeypatch):
+    def create_fake_distribution(name):
+        return Distribution(project_name="cerberus", version="1.2.3")
+
+    with monkeypatch.context() as m:
+        cerberus = __import__('cerberus')
+        m.setattr('pkg_resources.get_distribution', create_fake_distribution)
+        reload(cerberus)
+        assert cerberus.__version__ == '1.2.3'
+
+
+def test_version_not_found(monkeypatch):
+    def raise_distribution_not_found(name):
+        raise DistributionNotFound("pkg_resources cannot get distribution")
+
+    with monkeypatch.context() as m:
+        cerberus = __import__('cerberus')
+        m.setattr('pkg_resources.get_distribution', raise_distribution_not_found)
+        reload(cerberus)
+        assert cerberus.__version__ == 'unknown'
 
 
 def test_clear_cache(validator):
